Apply version switcher to get the correct serializable resource class
