+++ lib/google_json_response/version.rb
@@ -1,3 +1,3 @@
 module GoogleJsonResponse
-  VERSION = "0.1.4"
+  VERSION = "0.1.5"
 end
